requirements.txt
streamlit
chromadb-client
sentence-transformers
python-dotenv